Reorganise a bit to reflect which components are shared
