+++ support-frontend/assets/components/priceSummary/priceSummary.tsx
