+++ support-frontend/assets/pages/[countryGroupId]/components/checkoutForm.tsx
@@ -23,9 +23,9 @@ import { useEffect, useRef, useState } from 'react';
 import { Box, BoxContents } from 'components/checkoutBox/checkoutBox';
 import DirectDebitForm from 'components/directDebit/directDebitForm/directDebitForm';
 import { checkAccount } from 'components/directDebit/helpers/ajax';
-import { MaybeEstimatedTaxSummary } from 'components/orderSummary/maybeEstimatedTaxSummary';
 import { paymentMethodData } from 'components/paymentMethodSelector/paymentMethodData';
 import { Recaptcha } from 'components/recaptcha/recaptcha';
+import { MaybeEstimatedTaxSummary } from 'components/salesTax/maybeEstimatedTaxSummary';
 import { SecureTransactionIndicator } from 'components/secureTransactionIndicator/secureTransactionIndicator';
 import { StripeCardForm } from 'components/stripeCardForm/stripeCardForm';
 import type { AddressFormFieldError } from 'components/subscriptionCheckouts/address/addressFields';
